requirements.txt
appdirs==1.4.4
asgiref==3.3.4
distlib==0.3.2
Django==3.2.4
django-multipleselectfield==0.2.0.post2
django-multiselectfield==0.1.12
django-phonenumber-field==5.2.0
djangorestframework==3.12.4
filelock==3.0.12
phonenumbers==8.12.25
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
virtualenv==20.4.7
